fix check history
